7_Implementación_de_Rate_Limiting_en_un_Endpoint/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ratelimit==4.1.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
